+++ lib/elixir/lib/enum.ex
@@ -1191,7 +1191,8 @@ defmodule Enum do
   def group_by(enumerable, dict, fun) do
     IO.warn(
       "Enum.group_by/3 with a map/dictionary as second element is deprecated. " <>
-        "A map is used by default and it is no longer required to pass one to this function"
+        "A map is used by default and it is no longer required to pass one to this function",
+      Process.stacktrace(drop: 1)
     )
 
     # Avoid warnings about Dict