+++ lib/mix/lib/mix/compilers/erlang.ex
@@ -60,7 +60,8 @@ defmodule Mix.Compilers.Erlang do
       when is_boolean(force) or is_nil(force) do
     IO.warn(
       "Mix.Compilers.Erlang.compile/6 with a boolean or nil as 5th argument is deprecated, " <>
-        "please pass [force: true] or [] instead"
+        "please pass [force: true] or [] instead",
+      []
     )
 
     compile(manifest, mappings, src_ext, dest_ext, [force: force], callback)
@@ -240,7 +241,8 @@ defmodule Mix.Compilers.Erlang do
       {:ok, _} ->
         IO.warn(
           "returning {:ok, contents} in the Mix.Compilers.Erlang.compile/6 callback is deprecated " <>
-            "The callback should return {:ok, contents, warnings} or {:error, errors, warnings}"
+            "The callback should return {:ok, contents, warnings} or {:error, errors, warnings}",
+          []
         )
 
         {:ok, [], [], []}
@@ -248,7 +250,8 @@ defmodule Mix.Compilers.Erlang do
       :error ->
         IO.warn(
           "returning :error in the Mix.Compilers.Erlang.compile/6 callback is deprecated " <>
-            "The callback should return {:ok, contents, warnings} or {:error, errors, warnings}"
+            "The callback should return {:ok, contents, warnings} or {:error, errors, warnings}",
+          []
         )
 
         {:error, [], [], []}